Move SearchResultGeneration to org.kitodo.production.helper
